UI update - Add Dialog components to App. If user is not logged in, it show a dialog telling user to go back to the blog app to login before continue. - Remove other navigation option form the navigation bar in header except the logout option - Add loading screen to the CommentPage, EditPage, PostPage and DashboardPage
Minor fix and update
- Rename the prop of hidden to collasped in SideBar
- Change how the arguments are passed to the createPost and the updatePost
method of apiService. Arguments should now be passed separately to the
method instead of being passed as object. The EditPage ant the PostPage
were also updated to reflect the changes.

Add ThemeProvider
- Move the theme feature from ThemeFlyout to the ThemeProvider. The entire
app now can access the theme instead of limiting to the ThemeFlyout.

Update to reflect the change of error handling in the backend
- ensure all relevant files are updated to reflect the changes made to
the error handling system, the files include:
    -ProfileProvider
    -NotificationProvider
    -AuthenProvider
    -apiService